(PC-41797)[API] refactor: rename `cds_client` to `cine_office_client`
